Fix return code for --help and --version
When the `--help' or `--version' command-line options were given,
quicktun would exit with exit status 1, although the requested action
was fulfilled without encountering an error. With this change, these
options now will result in a zero exit status.

In addition, with both of these option, the header printed was partially
overlapping with the message printed by the option, so the header
is now printer after option processing.